@@ -3,7 +3,7 @@ description: A common platform interface for the audio_service plugin. Different
 homepage: https://github.com/ryanheise/audio_service/tree/master/audio_service_platform_interface
 # NOTE: We strongly prefer non-breaking changes, even at the expense of a
 # less-clean API. See https://flutter.dev/go/platform-interface-breaking-changes
-version: 0.1.0
+version: 0.1.1
 
 dependencies:
   flutter:
@@ -19,5 +19,5 @@ dev_dependencies:
   build_runner: ^2.0.4
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5"
+  sdk: ">=2.14.0 <4.0.0"
+  flutter: ">=3.0.0"
