@@ -4,13 +4,13 @@ version: 0.1.1
 homepage: https://github.com/ryanheise/audio_service/tree/master/audio_service_web
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=2.0.0"
+  sdk: ">=2.14.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   # Use these deps for local development
-  # audio_service_platform_interface:
-  #   path: ../audio_service_platform_interface
+  #audio_service_platform_interface:
+  #  path: ../audio_service_platform_interface
 
   # Use these deps when pushing to origin (for the benefit of testers)
   # audio_service_platform_interface:
@@ -20,7 +20,7 @@ dependencies:
   #     path: audio_service_platform_interface
 
   # Use these deps when publishing.
-  audio_service_platform_interface: ^0.1.0
+  audio_service_platform_interface: ^0.1.1
 
   rxdart: '>=0.26.0 <0.28.0'
   js: ^0.6.3
